+++ react/components/SearchQueryContainer.js
@@ -29,7 +29,6 @@ class SearchQueryContainer extends Component {
     const {
       treePath,
       params,
-      runtime,
       query: {
         order: orderBy = SortOptions[0].value,
         page: pageProps,
@@ -38,7 +37,7 @@ class SearchQueryContainer extends Component {
       },
     } = this.props
 
-    const path = reversePagesPath(runtime, treePath, params)
+    const path = reversePagesPath(treePath, params)
     const map = mapProps || createMap(path, rest)
     const page = pageProps ? parseInt(pageProps) : DEFAULT_PAGE
     const query = joinPathWithRest(path, rest)